% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rope.R
\name{rope}
\alias{rope}
\title{Robust profile likelihood-based differential expression analysis}
\usage{
rope(datmat, X_model, x_PI_idx)
}
\arguments{
\item{datmat}{A matrix of filtered RNA-seq counts.}

\item{X_model}{A numerical design matrix contains DE experiment group (parameter of interest) and other covariates}

\item{x_PI_idx}{An integer indicates the column index of the parameter of interest, usually the DE experiment group.}
}
\value{
A data frame with columns contain the DE result of log fold change, unadjusted and adjusted log likelihood ratio, adjustment factors, p-values and Benjamini & Hochberg adjusted p-values.
}
\description{
Fit a robust profile likelihood ratio test for a working Poisson log-linear model.
}
\examples{
count <- matrix(rnbinom(n = 5e4, mu = 100, size = 1 / 0.5), ncol = 50)
mod <- model.matrix(~ gl(n = 2, k = 25))
res <- rope(datmat = count, X_model = mod, x_PI_idx = dim(mod)[2])
res
}
